implementando outras exceções

+++ src/main/java/com/formacaospring/dscommerce/services/ProductService.java
@@ -1,13 +1,19 @@
 package com.formacaospring.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.formacaospring.dscommerce.repositories.ProductRepository;
+import com.formacaospring.dscommerce.services.exceptions.DatabaseException;
 import com.formacaospring.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
+
 import com.formacaospring.dscommerce.dto.ProductDTO;
 import com.formacaospring.dscommerce.entities.Product;
 
@@ -18,21 +24,20 @@ public class ProductService {
     private ProductRepository repository;
 
     @Transactional(readOnly = true)
-    public ProductDTO findById(Long id){
+    public ProductDTO findById(Long id) {
         Product product = repository.findById(id).orElseThrow(
-            () -> new ResourceNotFoundException("Recurso não encontrado!")
-        );
+                () -> new ResourceNotFoundException("Recurso não encontrado!"));
         return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
-    public Page<ProductDTO> findAll(Pageable pageable){
+    public Page<ProductDTO> findAll(Pageable pageable) {
         Page<Product> result = repository.findAll(pageable);
         return result.map(x -> new ProductDTO(x));
     }
 
     @Transactional
-    public ProductDTO insert(ProductDTO dto){
+    public ProductDTO insert(ProductDTO dto) {
         Product entity = new Product();
         copyDtoToentity(dto, entity);
         entity = repository.save(entity);
@@ -40,17 +45,30 @@ public class ProductService {
     }
 
     @Transactional
-    public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);       
-        copyDtoToentity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+    public ProductDTO update(Long id, ProductDTO dto) {
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToentity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+
     }
 
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
-        
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+        try{
+            repository.deleteById(id);
+        }
+        catch(DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial!");
+        }
+
     }
 
     private void copyDtoToentity(ProductDTO dto, Product entity) {
@@ -59,5 +77,5 @@ public class ProductService {
         entity.setPrice(dto.getPrice());
         entity.setImgUrl(dto.getImgUrl());
     }
-        
+
 }
